// src/lib/interfaces/ListaDeCliente.ts
export interface ListaDeCliente {
    razaosocial: string;
    cnpj: string;
    mesreferente: string;
    medicalLicencasAtivas: number;
    cooperadosAtivos: number;
    faturado: number;
    recebido: number;
    glosado: number;
    recebidoGlosa: number;
    quantidadeGlosa: number;
    versao: string;
    dataDeEnvio: string;
    folhaFechada: number;
    nomeFantasia: string;
    id: number;
  }